mcp-server-factory/requirements.txt
# MCP Server Factory Requirements

# Core Framework
fastmcp>=2.0.0,<3.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & Orchestration
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# Tracing & Observability
arize-phoenix>=8.0.0
openinference-instrumentation-langchain>=0.1.0
opentelemetry-sdk>=1.28.0

# HTTP Client
httpx>=0.27.0

# Template Engine
jinja2>=3.1.0

# YAML Storage
pyyaml>=6.0

# Environment
python-dotenv>=1.0.0
